Add functionality to report broken course links
A new function to report broken course links has been added in the courses file. This function checks user authentication and the existence of the course before setting the relationship document with the user ID and course ID. Adjustments were made to improve the code formatting and comments. Additionally, Axios was added as a dependency in package.json.